winding snake
Creates a growing segmented cylinder that randomly wanders through space
Every second a new segment is added to the cylinder. The current path
direction is randomly permuted. We then create a matrix to represent the
transform ito the local space at the end of the path and transform our
base vertices into this space. Triangles are then constructed from these
new vertices and the tail vertices from the last cylinder. The GL buffer
is then updated with the full vertex set.